patch_checker.sh was in the wrong place

+++ html/client/patch_checker.sh
@@ -22,6 +22,8 @@ if [[ -f /etc/lsb-release && -f /etc/debian_version ]]; then
         os=$(lsb_release -s -d|head -1|awk {'print $1'})
 elif [[ -f /etc/debian_version ]]; then
         os="$(cat /etc/issue|head -n 1|awk {'print $1'})"
+elif [[ -f /etc/oracle-release ]]; then
+		os="Oracle"
 elif [[ -f /etc/redhat-release ]]; then
         os=$(cat /etc/redhat-release|head -1|awk {'print $1'})
         if [[ "$os" = "Red" && $(grep -i enterprise /etc/redhat-release) != "" ]]; then
@@ -35,7 +37,53 @@ fi
 # remove any special characters
 os=$(echo $os|sed -e 's/[^a-zA-Z0-9]//g')
 # begin update checks
-if [[ "$os" = "CentOS" ]] || [[ "$os" = "Fedora" ]] || [[ "$os" = "RHEL" ]]; then
+if [[ "$os" = "Oracle" ]]; then
+	need_patched="true"
+
+	osver=`awk '{print $5}' /etc/oracle-release`
+
+	updateinfo="/tmp/updateinfo.txt"
+	checkupdate="/tmp/check-update.txt"
+	severityfile="/tmp/severity_sorted.txt"
+	
+	baseurl="https://linux.oracle.com/errata/"
+	
+	#Get list of all updates with severity
+	yum updateinfo list | awk '$1 ~/EL/ {print}' > $updateinfo
+
+	#Sort the list by severity
+	grep 'Critical/Sec.' $updateinfo | sort -r | sed 's:Critical/Sec.:Critical:g' > $severityfile
+	grep 'Important/Sec.' $updateinfo | sort -r | sed 's:Important/Sec.:Important:g' >> $severityfile
+	grep 'Moderate/Sec.' $updateinfo | sort -r | sed 's:Moderate/Sec.:Moderate:g' >> $severityfile
+	grep 'Low/Sec.' $updateinfo | sort -r | sed 's:Low/Sec.:Low:g' >> $severityfile
+	grep 'bugfix' $updateinfo | sort -r >> $severityfile
+	grep 'enhancement' $updateinfo | sort -r >> $severityfile
+
+	#Gets the list of the latest updates
+	yum -q check-update | awk '$3 ~/^ol/ {print}' > $checkupdate
+
+	while read line
+	do
+		pkg=`echo $line | awk -F. '{print $1}'`
+		instver=`rpm -q $pkg --qf '%{VERSION}-%{RELEASE}'`
+		
+		oldIFS=$IFS
+		IFS=$(echo -en "\n\b")
+		  		
+  		for updateline in `grep " $pkg-[0-9]" $severityfile`
+  		do
+  			updver=`echo $updateline | awk '{print $3}' | sed "s/$pkg-//g" | awk -F. 'sub(FS $NF,x)'`
+  			severity=`echo $updateline | awk '{print $2}'`
+  			advisory=`echo $updateline | awk '{print $1}'`
+  			  			
+			echo "$pkg:::$instver:::$updver:::$os:::$osver:::$severity:::$baseurl$advisory" >> /tmp/patch_$client_key
+  		done
+  		IFS=$oldIFS
+	done < $checkupdate
+
+	#Clean up text files
+	rm -rf $updateinfo $checkupdate $severityfile
+elif [[ "$os" = "CentOS" ]] || [[ "$os" = "Fedora" ]] || [[ "$os" = "RHEL" ]]; then
 	need_patched="true"
         yum -q check-update| while read i
         do
@@ -47,22 +95,25 @@ if [[ "$os" = "CentOS" ]] || [[ "$os" = "Fedora" ]] || [[ "$os" = "RHEL" ]]; the
                         PNAME=${i%%\ *}
                         PNAME=${PNAME%.*}
                         #echo $(rpm -q "${PNAME}" --qf '%{NAME}:::%{VERSION}:::')${UVERSION}
-                        patches_to_install=$(echo $(rpm -q "${PNAME}" --qf '%{NAME}:::%{VERSION}:::')${UVERSION})
+                        patches_to_install=$(echo $(rpm -q "${PNAME}" --qf '%{NAME}:::%{VERSION}-%{RELEASE}:::')${UVERSION})
                         echo "$patches_to_install" >> /tmp/patch_$client_key
                 fi
         done
 elif [[ "$os" = "Ubuntu" ]] || [[ "$os" = "Debian" ]]; then
         need_patched="true"
         #apt-get --just-print upgrade 2>&1 | perl -ne 'if (/Inst\s([\w,\-,\d,\.,~,:,\+]+)\s\[([\w,\-,\d,\.,~,:,\+]+)\]\s\(([\w,\-,\d,\.,~,:,\+]+)\)? /i) {print "$1:::$2:::$3\n"}'
+		apt-get -qq update
         patches_to_install=$(apt-get --just-print upgrade 2>&1 | perl -ne 'if (/Inst\s([\w,\-,\d,\.,~,:,\+]+)\s\[([\w,\-,\d,\.,~,:,\+]+)\]\s\(([\w,\-,\d,\.,~,:,\+]+)\)? /i) {print "$1:::$2:::$3\n"}')
+	extrainfo=$(lsb_release -s -d | awk '{gsub( "[\(\)]","" ); print ":::"$1":::"$4}')
 	echo "$patches_to_install" >> /tmp/patch_$client_key
+	sed -i 's/$/'$extrainfo'/g' /tmp/patch_$client_key
 elif [[ "$os" = "Linux" ]]; then
         echo "unspecified $os not supported"
         exit 0
 fi
-if [[ "$need_patched" == "true" ]]; then
+if [[ "$need_patched" == "true" ]] && [ -f /tmp/patch_$client_key ] ; then
         patch_list=$(cat /tmp/patch_$client_key)
-        curl -k -s -H "X-CLIENT-KEY: $client_key" $submit_patch_uri -d "$patch_list"
+        curl -s -H "X-CLIENT-KEY: $client_key" $submit_patch_uri -d "$patch_list"
         rm -rf /tmp/patch_$client_key > /dev/null 2>&1
 fi
